chore: introduce ty type checker (#122)
* chore: introduce ty type checker

Add ty (pinned, preview) as a dev dependency with a [tool.ty] config scoped to gitsvg/. Exclude render/ and theme/_theme.py, which concentrate the Optional-typed resolved-theme errors (the resolved RendererSettings still inherits the Theme base optional fields); ty checks the rest of the library green.

Fix the independent findings ty surfaced: widen ValidationError.file to str | None (legitimately None for errors with no source location), type a forwarded pydantic-error dict as Mapping, narrow op.name and a resolved orientation, cast the Click-validated minify level, and align an overridden _resolve_merge_commit_radius with its base. Add a make typecheck target.


* fix: render source-less validation errors without the None:0: prefix

Validation errors with no source location (e.g. a named-theme table layout clashing with a horizontal orientation) previously rendered a confusing None:0: prefix; ValidationError.format() now omits the location when file is None.

Also unify make lint to run ruff plus ty (dropping the separate typecheck target), and add clarifying comments to the op.name narrowing and the minify-level cast.


* test: cover the source-less ValidationError rendering path

Add a test for ValidationError.format() with file=None (the no-location branch the rendering fix added). Statement coverage hid the gap because the ternary line still executed via the file-bearing cases.


* chore: let ty infer its target version from requires-python

Drop the hardcoded [tool.ty.environment] python-version; ty reads requires-python (>=3.11) instead, so the target is single-sourced and stays at the supported floor even when run under a newer interpreter (verified: targets 3.11 under a 3.13 interpreter).


---------

+++ gitsvg/theme/themes/_gui.py
@@ -100,7 +100,7 @@ class GuiTheme(DefaultTheme):
         return 3
 
     @classmethod
-    def _resolve_merge_commit_radius(cls, commit_radius: int) -> int:
+    def _resolve_merge_commit_radius(cls, commit_radius: float) -> float:
         """Merge dots stay at `5`.
 
         Independent of the smaller ordinary-dot `commit_radius`, so merges
